Added a few new features and rearranged a few.
Created a separate script for the "Input field check" as this is useful
outside of just the single script. Added a folder upload part to the
Google Drive portion of the script that creates a folder before dumping
all the photo files into it. Added a few try/except statements to
(hopefully) catch issues with authentication that won't force starting
from the beginning. Uncovered a few bugs when certain fields are not
selected, or filled in too early; will need to add checks to prevent
crashes in the future.