Fixed a few bugs, added a file counter for upload.
File uploader window now includes a dynamic file counter
so that as the progress bar increases, there is also a numeric
way of tracking how soon the upload will be completed.
Found a few bugs during testing, and corrected them. One of the bugs
was a incorrect Boolean in the mass renamer, causing it to not function.
Corrected another bug related to the multi folder creator that would not
properly check if the folder name was already in the folder when the
case of the new folder was different.