Added several new features and added comments.
Added ability to add a permissions file to allow file and
folder sharing with chosen users. Added a new selection window
to decide if a single file upload is needed and if file/folder should
be shared. Removed some of the extraneous lines of code for
the radio selections, and cleaned up selection variables from the
radios to make it clearer what is happening in the code when each is
selected. Added comments and spaces to both of the main files for the
scripts.