Fixed a few bugs and added/removed features.
Fixed a bug that prevented the for loop in the multiple folder making
function from executing properly and creating multiple folders.
Fixed bugs related to clicking into file list windows before a folder
had been selected. Removed the no_photo_renamer function as it didn't
really have much of a point except to remove a button. Added a delete
button to the photo renamer function so that individual or groups of
files/photos can be deleted at will.