Expanded the sharing capabilities of File_uploader.
Added a way to extract the folder ID from local folders
and shared Drive folders so that the user can select the
folder to upload photos/.pdf into. This still requires adding a user
to a permissions.json file, but the next step will be to eliminate
the need for that file. Also fixed the single upload selection.
Single file will now upload the last selected file in the Photo_renamer
function and upload it based on sharing preferences. Currently,
the no share option only goes into main Drive.