Updated functions for file MIMEType for upload.
Previously, .pdf would upload to Drive, but it would only be visible
after downloading as the MIMEType was set to be Image/jpeg. This has
been updated so that .pdf is able to be previewed in Drive.
Also fixed a small bug when .pdf files were instead notated as .pdf
in the Photo_renamer functions.