Sharing now works for single & multiple files.
Added a new function to handle file sharing for single
and multiple files. Added ability for the local folder holding the files
to be uploaded and used for final location name (instead of having to
input a new name and have that folder created on Google Drive).
Fixed a couple minor bugs, such as forgetting to remove a no longer
required variable from a function in the Photo_renamer.py module.

In case I need to reference this in the future, Google Drive already
has files and folders inherit permissions from its parent folder, so
there wasn't any need to try and extract the permissions and apply them
to each individual folder or file. As a result, the permissions.json
file is no longer needed to attach permissions to folders or files.