Added multiple pages for folder search
Added a way for large amounts of folders to be searched and added to the
shared folder finder. The current limit is up to 99 extra pages, but
can be expanded by extending the range. Also changed 'ok' to 'upload'
for when the token.json file needs to be deleted, which was a change
from an earlier iteration of the program.