Multiple bug fixes, popup cleanup and folder store
Added a new "feature" where the Google Drive downloaded shared folder
list is saved and can be edited in the folder with the rest of the
program to help cut down on download time for large Drives. (.gitignore)

Removed the default popup titles for multiple files so that the popup
windows come up with a "clean" appearance instead of duplicating their
text output as the title. (File_uploader, Input_field_check,
Photo_renamer)

Fixed various bugs related to not selecting radios when prompted, which
wouldn't crash the program, but would create confusion when suddenly
a folder/files were being uploaded during the File_uploader.py script
(this folder/file would only be uploaded to the user's personal Drive
and not shared, as radios default to false values). Also fixed a bug
when a folder was not selected for file/folder upload but "Ok" was
pushed anyways. Now the program prompts for the user to select a folder
and doesn't hang afterwards. (File_uploader, Photo_renamer)