Minor typo and code fixes
Fixed some references from early code for "photos" instead of "files."
Changed the name of a few buttons. Added a collapsed section for the
file uploader as desired that will warn the user about existing Drive
list being out of date and that a new list will take a while to update.
(This was previously just a text blurb below the radios). Corrected a
few code inconsistencies involving functions that were no longer needed
and 2 variables assigned to one Boolean that would immediately flip to
the opposite Boolean. Fixed a bug when moving files between folders that
would cause the a single file to be displayed in a folder even if all
the files in that folder had been moved over. Now the folder on the left
will show as empty once all the files are moved out of it.