Added rotate feature, fixed some other windows.
Added the ability to rotate photos within the application, at 90deg
angles, either clockwise or counterclockwise. Added a warning text box
before deleting a file so that the user knows it will permanently
delete the file. Changed the properties test popup so that the cancel
button no longer exists. Added a change for theme when closing the
theme selection window. Previously it was reading the temp theme change
and using that for the rest of the program execution. The 'Checks'
change is attempting to fix a bug that I have only found can be
reproduced once a week, when the token refreshes. The program will
delete the 'token.json' like it is supposed to, but will then exit. My
guess is because there was no 'creds' returned from the function.
The program starts back again and creates a new 'token.json' as
expected, the first crash is just unexpected.