Code cleanup, features added, README updated
Added a new custom icon for the program to use (drawn by myself and I
will update for better readability). Performed a wide sweep over all the
code to fix inconsistencies, change different ways that things were
being run (thanks enumerate()!), made a change in the file mover so that
it would not rediscover all the files in a folder but will now only
remove the file(s) from one folder and add it(them) to the destination
folder. Normally the old method was fine, but if there are a lot of
files in one or both of the folders, it would take a while to update.
Added spaces in places, added some comments (more still needed), and
generally tried to make sure the style is consistent throughout.