Bug fix/window correction and splitting delete off
Fixed issue with multiple photo folder creation where the program would
exit after the first folder creation. The program was behaving
correctly, but obviously not the way it was intended. The issue comes
from how PySimpleGUI window.close() operates; it uses the same
operation for a WIN_CLOSED event to tear down a window. In the "bug"
case, the tear down was sys.exit(), which stopped the program. A break
like the one that is normally called for window closure works... but the
windows are created in a for loop which causes the current window to
exit... and start the next window. To get around this, another window
exit check was employed to cause the 'X' to close the program as it
should, and the for loop to continue making windows as it should too.
Created a new delete() function to enable future spots to call delete()
if needed. The one spot it is needed is in the upload folder area, as
a user should be able to edit the saved Drive folders in the program
instead of using another text editor or IDE to do so.